+++ lib/turnkit.rb
@@ -10,6 +10,7 @@ require_relative "turnkit/version"
 require_relative "turnkit/error"
 require_relative "turnkit/id"
 require_relative "turnkit/clock"
+require_relative "turnkit/cost"
 require_relative "turnkit/budget"
 require_relative "turnkit/agent"
 require_relative "turnkit/client"
@@ -42,6 +43,7 @@ module TurnKit
     attr_accessor :default_model, :client, :store, :logger
     attr_accessor :max_iterations, :timeout, :max_depth, :max_tool_executions
     attr_accessor :cost_limit, :prompt_cache
+    attr_accessor :cost_rates, :cost_calculator
     attr_accessor :prompt_sections, :prompt_behavior, :available_skills
     attr_accessor :prompt_data_max_chars, :context_contributors
     attr_accessor :system_prompt_contributors, :model_prompt_contributors
@@ -57,6 +59,7 @@ module TurnKit
   self.max_depth = 3
   self.max_tool_executions = 100
   self.prompt_cache = :auto
+  self.cost_rates = {}
   self.prompt_sections = SystemPrompt::DEFAULT_SECTIONS.dup
   self.prompt_data_max_chars = 20_000
   self.available_skills = []