+++ test/test_helper.rb
@@ -16,6 +16,8 @@ class Minitest::Test
     TurnKit.max_tool_executions = 100
     TurnKit.timeout = 300
     TurnKit.cost_limit = nil
+    TurnKit.cost_rates = {}
+    TurnKit.cost_calculator = nil
     TurnKit.prompt_cache = :auto
     TurnKit.prompt_sections = TurnKit::SystemPrompt::DEFAULT_SECTIONS.dup
     TurnKit.prompt_behavior = nil